requirements.txt
langchain
langchain-community
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain-groq